Move grid renderers to correct package
Only renderers introduced in 8.1 have been moved for
backwards compatibility.
